pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ocmfet_client"
version = "2.5.0"
description = "A client for the OCMFET acquisition system."
authors = [{ name = "Fabio Terranova" }]
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"numpy==1.26.4",
	"PyQt5==5.15.10",
	"matplotlib==3.10.0",
	"scikit-learn==1.6.1",
	"pyqtgraph==0.13.7",
	"PyYAML==6.0.1",
	"scipy==1.15.1",
	"oCPPmfet @ {root:uri}/src/oCPPmfet",
]

[project.scripts]
ocmfet_client = "ocmfet_client:cli"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.default]
dependencies = ["pre-commit"]

[tool.hatch.envs.build]
include = ["/src/configs/*.yaml"]
